refactor(frontend): split sidebar menu item into separate components
