+++ apps/frontend/src/app/components/molecules/index.tsx
@@ -10,3 +10,6 @@ export * from './chart/datePicker'
 export * from './chart/dateRangePicker'
 export * from './dragAndDrop/sortableItem'
 export * from './dragAndDrop/addCategoryDialog'
+export * from './sidebar/menuItem/app-normal-menuItem'
+export * from './sidebar/menuItem/app-collapsible-menuItem'
+export * from './sidebar/menuItem/app-popover-menuItem'
